server/requirements.txt
appnope==0.1.0
decorator==4.0.4
Django==1.8.5
django-cors-headers==1.1.0
django-filter==0.7
django-rest-auth==0.5.0
djangorestframework==3.2.4
djangorestframework-csv==1.3.4
docutils==0.12
easy-thumbnails==2.2
gnureadline==6.3.3
ipython==4.0.0
ipython-genutils==0.1.0
Jinja2==2.7.3
jsonfield==1.0.0
Markdown==2.5
MarkupSafe==0.23
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
Pillow==3.0.0
ptyprocess==0.5
Pygments==1.6
python-dateutil==2.4.0
pytz==2015.2
requests==2.8.1
simplegeneric==0.8.1
simplejson==3.6.3
six==1.8.0
traitlets==4.0.0
Werkzeug==0.9.6
wheel==0.24.0
